Refined newsletter management for logged-in members
refs https://github.com/TryGhost/Team/issues/1559

- extracts newsletter management in a common component for use across pages
